+++ internal/randvar/skewed_latest.go
@@ -52,7 +52,7 @@ func NewSkewedLatest(min, max uint64, theta float64) (*SkewedLatest, error) {
 }
 
 // IncMax increments max.
-func (z *SkewedLatest) IncMax(delta int) {
+func (z *SkewedLatest) IncMax(delta uint64) {
 	z.mu.Lock()
 	z.mu.zipf.IncMax(delta)
 	z.mu.max += uint64(delta)
